updated backend code
I added some new alerts which would be printed to the console when there are some specific errors

+++ Backend/controllers/product.controller.js
@@ -48,14 +48,18 @@ export const updateProduct = async(req, res) => {
 
 export const deleteProduct = async(req, res) => {
     const {id} = req.params; // get the id from the url
+
+    // check if the product id exists or not
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: 'Invalid Product Id' });
+    }
     
     try {
         await Product.findByIdAndDelete(id); // find the product by id and delete it
         res.status(200).json({ success: true, message: 'Product deleted successfully' });
     } 
     catch (error) {
-        console.error("Error in deleting product:", error.message); 
-        res.status(404).json({ success: false, message: 'Product not found' });   
-        //return res.status(500).json({ success: false, message: 'Server Error' });
+        console.error("Error in deleting product:", error.message);   
+        res.status(500).json({ success: false, message: 'Server Error' });
     }
 }
